+++ app/components/Navbar/Navbar.php
@@ -7,6 +7,7 @@ use App\Constants\Container\SystemGroups;
 use App\Constants\ContainerStatus;
 use App\Core\Caching\CacheFactory;
 use App\Core\Caching\CacheNames;
+use App\Core\FileManager;
 use App\Core\Http\HttpRequest;
 use App\Entities\UserEntity;
 use App\Managers\Container\GroupManager;
@@ -166,12 +167,30 @@ class Navbar extends AComponent {
             default:
                 break;
         }
+        
+        if($this->user->getProfilePictureFileId() !== null) {
+            $file = $this->app->fileStorageManager->getFileById($this->user->getProfilePictureFileId());
 
-        if($link === null) {
-            return '<span class="navbar-link" style="cursor: pointer" title="' . $this->user->getFullname() . '">' . $this->user->getFullname() . '</span>';
+            $imageSource = FileManager::getRelativeFilePathFromAbsoluteFilePath($file->filepath);
+        } else {
+            $imageSource = 'resources/images/user-profile-picture.png';
         }
 
-        return $this->createLink($link, $this->user->getFullname());
+        $profileImage = '
+            <img src="' . $imageSource . '" width="24px" height="24px" style="border-radius: 100px">
+        ';
+
+        $content = $profileImage . '&nbsp;&nbsp;' . $this->user->getFullname();
+
+        if($link === null) {
+            return '
+                <span class="navbar-link" style="cursor: pointer" title="' . $this->user->getFullname() . '">
+                    ' . $content . '
+                </span>
+            ';
+        } else {
+            return $this->createLink($link, $this->user->getFullname(), $content);
+        }
     }
 
     /**
@@ -208,10 +227,14 @@ class Navbar extends AComponent {
      * 
      * @param array $url Link URL
      * @param string $title Link title
+     * @param ?string $content Link content
      * @return string HTML code
      */
-    private function createLink(array $url, string $title) {
-        return '<a class="navbar-link" href="' . LinkBuilder::convertUrlArrayToString($url) . '" title="' . $title . '">' . $title . '</a>';
+    private function createLink(array $url, string $title, ?string $content = null) {
+        if($content === null) {
+            $content = $title;
+        }
+        return '<a class="navbar-link" href="' . LinkBuilder::convertUrlArrayToString($url) . '" title="' . $title . '">' . $content . '</a>';
     }
 
     public function render() {