+++ app/components/Widgets/FileStorageStatsWidget/FileStorageStatsWidget.php
@@ -3,14 +3,8 @@
 namespace App\Components\Widgets\FileStorageStatsWidget;
 
 use App\Components\Widgets\Widget;
-use App\Core\DB\DatabaseManager;
-use App\Core\Http\HttpRequest;
 use App\Exceptions\AException;
-use App\Exceptions\GeneralException;
 use App\Helpers\UnitConversionHelper;
-use App\Logger\Logger;
-use App\Managers\ContainerManager;
-use App\Repositories\Container\FileStorageRepository;
 
 /**
  * FileStorageStatsWidget display statistics for application file storage
@@ -18,31 +12,6 @@ use App\Repositories\Container\FileStorageRepository;
  * @author Lukas Velek
  */
 class FileStorageStatsWidget extends Widget {
-    private ContainerManager $containerManager;
-    private DatabaseManager $dbManager;
-    private Logger $logger;
-
-    /**
-     * Class constructor
-     * 
-     * @param HttpRequest $request
-     * @param ContainerManager $containerManager
-     * @param DatabaseManager $dbManageer
-     * @param Logger $logger
-     */
-    public function __construct(
-        HttpRequest $request,
-        ContainerManager $containerManager,
-        DatabaseManager $dbManager,
-        Logger $logger
-    ) {
-        parent::__construct($request);
-
-        $this->containerManager = $containerManager;
-        $this->dbManager = $dbManager;
-        $this->logger = $logger;
-    }
-
     public function startup() {
         parent::startup();
 
@@ -59,8 +28,7 @@ class FileStorageStatsWidget extends Widget {
 
         $rows = [
             'All files' => $data['totalRows'],
-            'Total file size' => $data['totalSize'],
-            'Average file count for containers' => $data['avgFileCountForContainers']
+            'Total file size' => $data['totalSize']
         ];
 
         return $rows;
@@ -72,132 +40,46 @@ class FileStorageStatsWidget extends Widget {
     private function fetchDataFromDb(): array {
         $totalFiles = 0.0;
         $totalFileSize = 0.0;
-        $avgFileCountForContainers = 0.0;
         
         try {
             $totalFiles = $this->fetchTotalFileCountFromDb();
             $totalFileSize = $this->fetchTotalFileSizeFromDb();
-            $avgFileCountForContainers = $this->fetchAverageFileCountForContainersFromDb();
-        } catch(AException $e) {
-            
-        }
+        } catch(AException $e) {}
 
         return [
             'totalRows' => $totalFiles,
-            'totalSize' => $totalFileSize,
-            'avgFileCountForContainers' => $avgFileCountForContainers
+            'totalSize' => $totalFileSize
         ];
-    }
-
-    /**
-     * Fetches average file count for container from database
-     */
-    private function fetchAverageFileCountForContainersFromDb(): int {
-        $containers = $this->getAllContainers();
-
-        $storedFiles = 0;
-        foreach($containers as $containerId => $container) {
-            try {
-                $conn = $this->dbManager->getConnectionToDatabase($container->getDefaultDatabase()->getName());
-            } catch(AException $e) {
-                throw new GeneralException('Could not establish connection to database for container ID \'' . $containerId . '\'.', $e);
-            }
-
-            $fileStorageRepository = new FileStorageRepository($conn, $this->logger, $this->app->transactionLogRepository);
-            $fileStorageRepository->setContainerId($containerId);
-
-            $qb = $fileStorageRepository->composeQueryForStoredFiles();
-            $qb->select(['COUNT(*) AS cnt'])
-                ->regenerateSQL()
-                ->execute();
-
-            while($row = $qb->fetchAssoc()) {
-                $storedFiles += (int)($row['cnt']);
-            }
-        }
-
-        if(count($containers) == 0) {
-            return 0;
-        } else {
-            return ceil($storedFiles / count($containers));
-        }
     }
 
     /**
      * Fetches total file count from database
      */
     private function fetchTotalFileCountFromDb(): int {
-        $containers = $this->getAllContainers();
+        $qb = $this->app->fileStorageRepository->composeQueryForFilesInStorage();
+        $qb->select(['COUNT(*) AS cnt'])
+            ->regenerateSQL()
+            ->execute();
 
-        $storedFiles = 0;
-        foreach($containers as $containerId => $container) {
-            try {
-                $conn = $this->dbManager->getConnectionToDatabase($container->getDefaultDatabase()->getName());
-            } catch(AException $e) {
-                throw new GeneralException('Could not establish connection to database for container ID \'' . $containerId . '\'.', $e);
-            }
-
-            $fileStorageRepository = new FileStorageRepository($conn, $this->logger, $this->app->transactionLogRepository);
-            $fileStorageRepository->setContainerId($containerId);
-
-            $qb = $fileStorageRepository->composeQueryForStoredFiles();
-            $qb->select(['COUNT(*) AS cnt'])
-                ->regenerateSQL()
-                ->execute();
-
-            while($row = $qb->fetchAssoc()) {
-                $storedFiles += (int)($row['cnt']);
-            }
-        }
-
-        return $storedFiles;
+        return $qb->fetch('cnt') ?? 0;
     }
 
     /**
      * Fetches total file size from database
      */
     private function fetchTotalFileSizeFromDb(): string {
-        $containers = $this->getAllContainers();
+        $qb = $this->app->fileStorageRepository->composeQueryForFilesInStorage();
+        $qb->select(['SUM(filesize) AS totalSize'])
+            ->regenerateSQL()
+            ->execute();
 
-        $filesize = 0;
-        foreach($containers as $containerId => $container) {
-            try {
-                $conn = $this->dbManager->getConnectionToDatabase($container->getDefaultDatabase()->getName());
-            } catch(AException $e) {
-                throw new GeneralException('Could not establish connection to database for container ID \'' . $containerId . '\'.', $e);
-            }
+        $result = $qb->fetch('totalSize');
 
-            $fileStorageRepository = new FileStorageRepository($conn, $this->logger, $this->app->transactionLogRepository);
-            $fileStorageRepository->setContainerId($containerId);
-
-            $qb = $fileStorageRepository->composeQueryForStoredFiles();
-            $qb->execute();
-
-            while($row = $qb->fetchAssoc()) {
-                $filesize += (int)($row['filesize']);
-            }
+        if($result !== null) {
+            return UnitConversionHelper::convertBytesToUserFriendly($result);
         }
 
-        return UnitConversionHelper::convertBytesToUserFriendly($filesize);
-    }
-
-    /**
-     * Returns an array with all containers
-     * 
-     * @return array<string, \App\Entities\ContainerEntity>
-     */
-    private function getAllContainers(): array {
-        $_containers = $this->containerManager->getAllContainers(true, true);
-
-        $containers = [];
-        foreach($_containers as $container) {
-            /**
-             * @var \App\Entities\ContainerEntity $container
-             */
-            $containers[$container->getId()] = $container;
-        }
-
-        return $containers;
+        return '0 B';
     }
 }