+++ app/components/static/UserProfileStatic/UserProfileStatic.php
@@ -2,6 +2,7 @@
 
 namespace App\Components\Static\UserProfileStatic;
 
+use App\Core\FileManager;
 use App\Core\Http\HttpRequest;
 use App\Entities\UserEntity;
 use App\Exceptions\GeneralException;
@@ -10,6 +11,7 @@ use App\Managers\UserAbsenceManager;
 use App\Managers\UserManager;
 use App\Managers\UserSubstituteManager;
 use App\UI\AComponent;
+use App\UI\LinkBuilder;
 
 /**
  * UserProfileStatic component is used to display user's profile
@@ -59,6 +61,8 @@ class UserProfileStatic extends AComponent {
         $template = $this->getTemplate(__DIR__ . '\\template.html');
 
         $template->user_profile = $this->build();
+        $template->user_profile_picture = $this->getProfilePicture();
+        $template->user_profile_picture_change_link = $this->getProfilePictureChangeLink();
 
         return $template->render()->getRenderedContent();
     }
@@ -133,6 +137,51 @@ class UserProfileStatic extends AComponent {
         $superior = $this->userManager->getUserById($superiorId);
 
         return $superior->getFullname();
+    }
+
+    /**
+     * Returns user's profile picture
+     */
+    private function getProfilePicture(): string {
+        if($this->user->getProfilePictureFileId() !== null) {
+            $fileId = $this->user->getProfilePictureFileId();
+
+            $file = $this->app->fileStorageManager->getFileById($fileId);
+
+            $imageSource = FileManager::getRelativeFilePathFromAbsoluteFilePath($file->filepath);
+        } else {
+            $imageSource = 'resources/images/user-profile-picture.png';
+        }
+
+        return '<img src="' . $imageSource . '" width="128px" height="128px" style="border-radius: 100px">';
+    }
+
+    /**
+     * Returns profile picture change link
+     */
+    private function getProfilePictureChangeLink(): string {
+        if($this->app->currentUser->getId() != $this->user->getId()) {
+            return '';
+        }
+
+        return LinkBuilder::createSimpleLink(
+            'Change profile picture',
+            $this->createFullURL(
+                sprintf(
+                    '%s:%s',
+                    $this->presenter->moduleName,
+                    substr(
+                        $this->presenter->name,
+                        0,
+                        (
+                            strlen($this->presenter->name) - strlen('Presenter')
+                        )
+                    )
+                ),
+                'changeProfilePictureForm'
+            ),
+            'link'
+        );
     }
 
     public static function createFromComponent(AComponent $component) {}