+++ CHANGELOG.md
@@ -19,10 +19,13 @@ Previously changelog for versions `1.0`-`1.4` has been located in README.md. How
 - Added container technical account creation from Superadministration
 - Added support for user editing
 - Added support for password changing
+- Added global file storage
+- Added support for user profile pictures
 - Reworked user administration UI in containers
 - Fixed a bug where custom container processes are visible in container settings in superadministration
 - Removed support for container environments
     - Because SkyDocu itself is split to environments (PROD, TEST, DEV)
+- Removed per-container file storage
 
 ## SkyDocu `1.6` (06/21/2025)
 - Global transaction log